install Ros Noetic on Raspberry 4B (3B+)
The "Raspberry_ROS_Noetic_Install_.sh" script is an automated tool for installing the ROS Noetic version on a Raspberry Pi. It is primarily used in the "Embedded Real-Time Operating System for System-on-Chip" course to assist students who are falling behind in their progress or need to rebuild the environment due to failed experiments or system crashes.